requirements.txt
fastapi>=0.115.3
uvicorn>=0.31.1
google-auth>=2.23.4
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.108.0
pytz>=2024.2
python-dateutil>=2.9.0
pydantic>=2.11.7,<3.0.0
pymongo>=4.6.0
python-dotenv>=1.1.1
